Cleanup and fixes for l5

+++ zeropilot4.0/src/attitude_manager/attitude_manager.cpp
@@ -60,7 +60,7 @@ void AttitudeManager::amUpdate() {
     RawImuBatch_t imuData = imuDriver->readRawData();
     ScaledImuBatch_t scaledImuData = imuDriver->scaleIMUData(imuData);
     for (int i = 0; i < scaledImuData.count; i++) {
-        uint32_t deltaTicks = scaledImuData.data[i].timestamp - lastTimestamp; 
+        uint16_t deltaTicks = scaledImuData.data[i].timestamp - lastTimestamp; 
         lastTimestamp = scaledImuData.data[i].timestamp;
 
         // Make lastTimestamp hold a real timestamp the first iteration
